pyproject.toml
[tool.poetry]
name = "backend-iot"
version = "0.1.0"
description = ""
authors = ["Andronaco Marco <marco.andronaco@olivetti.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.3.3"
peewee = "^3.16.3"
requests = "^2.31.0"
python-dotenv = "^1.0.0"


[tool.poetry.group.prod.dependencies]
waitress = "^2.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
